New feature - pass/fail build based on snapshot diffs
This new optional feature will return an error code if the percy snapshots don't match the approved snapshots percy has. This is useful as a continuous or presubmit build check to prevent submissions if snapshots don't match.

Note that it does require enabling read access to builds from the Percy team. Contact support at Percy.io for help enabling this for your builds.